requirements.txt
ansible==2.9.25
asn1crypto==0.24.0
attrs==21.2.0
Automat==0.6.0
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
bidict==0.21.3
certifi==2018.1.18
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.6
click==6.7
colorama==0.3.7
ConfigArgParse==0.15.2
configobj==5.0.6
constantly==15.1.0
contextvars==2.4
cryptography==35.0.0
deepdiff==5.5.0
Deprecated==1.2.13
distlib==0.3.3
distro-info===0.18ubuntu0.18.04.1
dnspython==2.1.0
filelock==3.3.0
future==0.18.2
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
immutables==0.16
importlib-metadata==4.8.1
importlib-resources==5.2.2
incremental==16.10.1
ipaddress==1.0.23
Jinja2==3.0.2
jmespath==0.10.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
MarkupSafe==2.0.1
netaddr==0.8.0
netconan==0.12.2
netifaces==0.10.4
numpy==1.19.5
ordered-set==4.0.2
PAM==0.4.2
pandas==1.1.5
paramiko==2.7.2
passlib==1.7.4
pathspec==0.5.5
platformdirs==2.4.0
ply==3.11
pyasn1==0.4.2
pyasn1-modules==0.2.1
pybatfish @ git+https://github.com/batfish/pybatfish.git@741ffc310187e1e17ae6757861de651e3d903991
pycparser==2.18
pycrypto==2.6.1
PyGObject==3.26.1
PyNaCl==1.4.0
pynetbox==6.1.3
pyOpenSSL==17.5.0
pyserial==3.4
python-apt==1.6.4
python-dateutil==2.8.2
python-debian==0.1.32
pytz==2021.3
pyxdg==0.25
PyYAML==3.12
requests==2.26.0
requests-toolbelt==0.9.1
requests-unixsocket==0.1.5
SecretStorage==2.3.1
service-identity==16.0.0
simplejson==3.17.5
six==1.11.0
ssh-import-id==5.7
systemd-python==234
tabulate==0.8.9
textfsm==1.1.2
Twisted==17.9.0
typing-extensions==3.10.0.2
ufw==0.36
urllib3==1.22
virtualenv==20.8.1
wrapt==1.13.1
yamllint==1.10.0
zipp==3.6.0
zope.interface==4.3.2
